Add create-app to e2e and move common functionality into seperate modules
